Refer "module" field to esm-compat
If I have interpreted the package.json fields conventions jungle correctly, the "module" field would a way to point to an ES2015-based module but in all other aspects, only containing ES5-compatible JS code.

Found the issue as I integrated this module into dexie as a devDependencie, using rollup to include it. Rollup is run after transpilation and when it finds the "module" field, it integrates that module into my final ES5 code. When the CI tests sends this to browserstack for IE11 testing - it fails.

I don't know why we're still maintaining dexie for IE11 though in 2021, but, I'm just in a minor release so I need to stick to it.